Move channels to separate files and use default export for actions
