+++ integrations/instagram/src/actions/index.ts
@@ -1,2 +1,8 @@
-export * from './proactive-conversation'
-export * from './proactive-user'
+import { getOrCreateConversationDm } from './proactive-conversation'
+import { getOrCreateUser } from './proactive-user'
+import * as bp from '.botpress'
+
+export default {
+  getOrCreateUser,
+  getOrCreateConversationDm,
+} satisfies bp.IntegrationProps['actions']
