+++ integrations/instagram/src/setup.ts
@@ -1,0 +1,4 @@
+import * as bp from '.botpress'
+
+export const register: bp.IntegrationProps['register'] = async () => {}
+export const unregister: bp.IntegrationProps['unregister'] = async () => {}
